fix: 修复排行榜获取不到数据的问题
- 补充 openDataContext/index.json 配置文件（开放数据域必需）
- 修复 hiddenScore 用 parseInt 导致精度丢失的问题

+++ openDataContext/index.js
@@ -36,7 +36,7 @@ function fetchFriendData() {
           avatarUrl: item.avatarUrl || '',
           numbersFound: nf ? parseInt(nf.value) : 0,
           time: td ? parseFloat(td.value) : 0,
-          hiddenScore: hs ? parseInt(hs.value) : 0
+          hiddenScore: hs ? parseFloat(hs.value) : 0
         };
       });
